Bug fix: AgreementActivity always popup if a device has 2+ users
If multiple users logged on the same device, the 2nd and later user
will always see AgreementActivity.
The reason is because the SharedPreference didn't save multiple
users successfully.
We should not modify the Set instance returned by getStringSet(),
we should always make a new Set to update the String Set values
in SharedPreference.

Change-Id: Ia90c58c0d3c5f9cf98b3931dabdbd5ff24e94349