Fix the bug in file download.
The previous implementation is again a misuse of the video api callback
data assignment.

The file download is an AsyncTask, and the callbacks are passed from the
VideoViewFragment.

The image thumbnail of the video is also an AsyncTask, no need to use
the file download for such small size url fetch.

Change-Id: Id0cb3c9dde2aed570c79e5b8aa5118c06d4193d0